+++ packages/enraiged/src/Http/Controllers/Accounts/Edit.php
@@ -4,6 +4,7 @@ namespace Enraiged\Http\Controllers\Accounts;
 
 use App\Http\Controllers\Controller;
 use Enraiged\Accounts\Models\Account;
+use Enraiged\Accounts\Resources\AccountResource;
 use Enraiged\Http\Requests\Accounts\UpdateRequest;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
@@ -11,6 +12,12 @@ use Illuminate\Http\Request;
 class Edit extends Controller
 {
     use AuthorizesRequests;
+
+    /** @var  Account  The account model. */
+    protected $account;
+
+    /** @var  User  The request user. */
+    protected $user;
 
     /**
      *  Display the edit account form component.
@@ -20,16 +27,35 @@ class Edit extends Controller
      */
     public function __invoke(Request $request, Account $account = null)
     {
-        $account = preg_match('/^my\.account/', $request->route()->getName())
+        $this->account = preg_match('/^my\.account/', $request->route()->getName())
             ? $request->user()->account
             : $account;
+        $this->user = $request->user();
 
-        $this->authorize('edit', $account);
+        $this->authorize('edit', $this->account);
 
         $request = UpdateRequest::createFrom($request);
 
-        $template = $request->form()->edit($account);
+        $template = $request->form()->edit($this->account);
 
-        return inertia('accounts/Edit', ['builder' => $template]);
+        return inertia('accounts/Edit', [
+            'account' => AccountResource::from($this->account),
+            'builder' => $template,
+            'messages' => $this->messages(),
+        ]);
+    }
+
+    /**
+     *  Construct and return an array of the available page messages.
+     *
+     *  @return array
+     */
+    private function messages()
+    {
+        if (!$this->account->isMyself() && $this->user->account->isAdministrator()) {
+            return [message('You are updating these account details as an administrator.', 'warn')];
+        }
+
+        return [];
     }
 }