+++ resources/js/components/accounts/forms/AccountForm.vue
@@ -1,20 +1,25 @@
 <template>
     <primevue-card>
+        <template #header>
+            <header class="border-bluegray-100 border-bottom-1 p-3 surface-200">
+                <h3 class="auto-margin max-width-sm">
+                    This form will allow you to manage account and profile details.
+                </h3>
+            </header>
+        </template>
         <template #content>
             <vue-form class="adjacent-labels auto-margin max-width-sm" ref="accountForm"
                 :builder="builder"
                 @form:ready="ready = true">
-                <div v-if="ready">
-                    <administrative-options v-if="administrativeOptions"
-                        :form="accountForm"
-                        :section="administrativeOptions"/>
-                    <personal-information
-                        :form="accountForm"
-                        :section="personalInformation"/>
-                    <authentication-details
-                        :form="accountForm"
-                        :section="authenticationDetails"/>
-                </div>
+                <administrative-options v-if="administrativeOptions"
+                    :form="accountForm"
+                    :section="administrativeOptions"/>
+                <personal-information
+                    :form="accountForm"
+                    :section="personalInformation"/>
+                <authentication-details
+                    :form="accountForm"
+                    :section="authenticationDetails"/>
             </vue-form>
         </template>
     </primevue-card>