+++ resources/js/pages/accounts/Edit.vue
@@ -1,7 +1,8 @@
 <template>
     <main class="content main">
-        <page-header back-button :title="title"/>
+        <page-header back-button :header="header" :title="title"/>
         <section class="auto-margin container max-width-lg">
+            <page-messages class="mb-3" :messages="messages" @dismiss="messages.splice($event, 1)"/>
             <account-form :builder="builder"/>
         </section>
     </main>
@@ -11,6 +12,7 @@
 import AppLayout from '@/layouts/App.vue';
 import AccountForm from '@/components/accounts/forms/AccountForm';
 import PageHeader from '@/components/ui/pages/PageHeader.vue';
+import PageMessages from '@/components/ui/pages/PageMessages.vue';
 import PrimevueButton from 'primevue/button';
 
 export default {
@@ -19,22 +21,31 @@ export default {
     components: {
         AccountForm,
         PageHeader,
+        PageMessages,
         PrimevueButton,
     },
 
     props: {
+        account: {
+            type: Object,
+            required: true,
+        },
         builder: {
             type: Object,
             required: true,
         },
+        messages: {
+            type: Array,
+            default: [],
+        },
     },
 
     computed: {
-        myself() {
-            return this.$attrs.auth.user.id === this.builder.resource.id;
+        header() {
+            return this.account.is_myself ? 'Update My Account' : `Update Account: ${this.account.profile.name}`;
         },
         title() {
-            return this.myself ? 'Update My Account' : 'Update Account';
+            return this.account.is_myself ? this.header : 'Update Account';
         },
     },
 };